chore: students area - feature completed, pages to show, edit, delete and create students create separate controllers, students.js and teachers.js to control the specific routes belonging to its areas and to be more easy to read, add new functionatilies and fix in the future create functions on script.js - to change background-color in the body tag if the users is in the students area
