+++ data.json
@@ -2,19 +2,6 @@
   "teachers": [
     {
       "id": 1,
-      "avatar_url": "https://images.unsplash.com/photo-1548449112-96a38a643324?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80",
-      "name": "Paulo farias",
-      "birth": 748656000000,
-      "education": "mestrado",
-      "services": "Presencial",
-      "skills": [
-        "Português",
-        " Espanhol"
-      ],
-      "created_at": 1592786261290
-    },
-    {
-      "id": 2,
       "avatar_url": "https://images.unsplash.com/flagged/photo-1559475555-b26777ed3ab4?ixlib=rb-1.2.1&auto=format&fit=crop&w=675&q=80",
       "name": "Cristiano silva",
       "birth": 681350400000,
@@ -24,19 +11,34 @@
       "created_at": 1592786261290
     },
     {
-      "id": 3,
+      "id": 2,
       "avatar_url": "https://images.unsplash.com/photo-1507398941214-572c25f4b1dc?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1266&q=80",
       "name": "Paula junior",
       "birth": 711158400000,
       "education": "medio",
-      "services": "Ead",
-      "skills": [
-        "Russo",
-        " Italiano",
-        " Português"
-      ],
+      "services": "Presencial",
+      "skills": "Russo, Italiano",
       "created_at": 1595205364145
     }
   ],
-  "students": []
+  "students": [
+    {
+      "avatar_url": "https://images.unsplash.com/photo-1496317899792-9d7dbcd928a1?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=634&q=80",
+      "name": "Lais silva",
+      "birth": 1004918400000,
+      "email": "laissilva@gmail.com",
+      "grade": "3EM",
+      "duration": "40",
+      "id": 1
+    },
+    {
+      "avatar_url": "https://images.unsplash.com/photo-1483721310020-03333e577078?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=1400&q=80",
+      "name": "Yiago rodrigues",
+      "birth": 806371200000,
+      "email": "yiago@gmail.com",
+      "grade": "2EM",
+      "duration": "40",
+      "id": 2
+    }
+  ]
 }
