+++ views/layout.njk
@@ -11,12 +11,13 @@
     <body>
         <header>
             <div class="links">
-                <a href="/" class="active">Teachers</a>
-                <a href="#">Students</a>
+                <a href="/teachers">Teachers</a>
+                <a href="/students">Students</a>
             </div>
         </header>
 
         {% block content %}{% endblock %}
-        
+
+        <script src="/scripts.js"></script>        
     </body>
 </html>
